move the layer implementation to the index
